+++ pybdsim/Data.py
@@ -10,6 +10,7 @@ Data - read various output files
 from . import Constants as _Constants
 from . import _General
 
+from collections import defaultdict as _defaultdict
 import copy as _copy
 import glob as _glob
 import numpy as _np
@@ -181,6 +182,8 @@ def _LoadRoot(filepath):
         d = _ROOT.DataLoader(filepath)
         d.model = GetModelForPlotting(d) # attach BDSAsciiData instance for convenience
         d.header = Header(HeaderTree=d.GetHeaderTree())
+        if d.header.nOriginalEvents == 0:
+            d.header.nOriginalEvents = int(d.GetEventTree().GetEntries())
         return d
     elif fileType == "REBDSIM":
         print('REBDSIM analysis file - using RebdsimFile')
@@ -275,11 +278,13 @@ class Header(object):
         self.geant4Version = ""
         self.rootVersion   = ""
         self.clhepVersion  = ""
+        self.timeStamp     = ""
         self.fileType      = ""
         self.dataVersion   = -1
         self.analysedFiles = []
         self.combinedFiles = []
         self.trajectoryFilters = []
+        self.skimmedFile   = False
         self.nOriginalEvents = 0
         if 'TFile' in kwargs:
             self._FillFromTFile(kwargs['TFile'])
@@ -307,12 +312,42 @@ class Header(object):
         self.geant4Version = str(hi.geant4Version)
         self.rootVersion   = str(hi.rootVersion)
         self.clhepVersion  = str(hi.clhepVersion)
+        self.timeStamp     = str(hi.timeStamp).strip()
         self.fileType      = str(hi.fileType)
         self.dataVersion   = int(hi.dataVersion)
         self.analysedFiles = [str(s) for s in hi.analysedFiles]
         self.combinedFiles = [str(s) for s in hi.combinedFiles]
         self.trajectoryFilters = [str(s) for s in hi.trajectoryFilters]
+        self.skimmedFile   = bool(hi.skimmedFile)
         self.nOriginalEvents = int(hi.nOriginalEvents)
+
+class Spectra(object):
+    def __init__(self, nameIn=None):
+        self.name = nameIn
+        self.histograms = {}
+        self.histogramspy = {}
+        self.pdgids = set()
+        self.pdgidsSorted = []
+
+    def append(self, pdgid, hist, path, nameIn=None):
+        if nameIn:
+            self.name = nameIn
+        self.histograms[pdgid] = hist
+        self.histogramspy[pdgid] = TH1(hist)
+        self.pdgids.add(pdgid)
+        self._generateSortedList()
+
+    def _generateSortedList(self):
+        integrals = {pdgid:h.integral for pdgid,h in self.histogramspy.items()}
+        integralsSorted = sorted(integrals.items(), key=lambda item: item[1])
+        self.pdgidsSorted = [pdgid for pdgid,_ in sorted(integrals.items(), key=lambda item: item[1], reverse=True)]
+
+def ParseSpectraName(hname):
+    hn = hname.replace('Top_','')
+    hn = hn.replace('Spectra_','')
+    name,nth,pdgid = hn.split('_')
+    pdgid = int(pdgid)
+    return name+"_"+nth,pdgid
 
 class RebdsimFile(object):
     """
@@ -332,13 +367,20 @@ class RebdsimFile(object):
         self.filename = filename
         self._f = _ROOT.TFile(filename)
         self.header = Header(TFile=self._f)
-        self.histograms   = {}
+        self.histograms = {}
         self.histograms1d = {}
         self.histograms2d = {}
         self.histograms3d = {}
+        self.histogramspy = {}
+        self.histograms1dpy = {}
+        self.histograms2dpy = {}
+        self.histograms3dpy = {}
+        self.spectra = _defaultdict(Spectra)
         self._Map("", self._f)
         if convert:
             self.ConvertToPybdsimHistograms()
+
+        self._PopulateSpectraDictionaries()
 
         def _prepare_data(branches, treedata):
             data = BDSAsciiData()
@@ -424,10 +466,6 @@ class RebdsimFile(object):
         """
         Convert all root histograms into numpy arrays.
         """
-        self.histogramspy = {}
-        self.histograms1dpy = {}
-        self.histograms2dpy = {}
-        self.histograms3dpy = {}
         for path,hist in self.histograms1d.items():
             hpy = TH1(hist)
             self.histograms1dpy[path] = hpy
@@ -441,6 +479,42 @@ class RebdsimFile(object):
             self.histograms3dpy[path] = hpy
             self.histogramspy[path] = hpy
 
+    def _PopulateSpectraDictionaries(self):
+        for path,hist in self.histograms1d.items():
+            hname = str(hist.GetName())
+            if 'Spectra' in hname:
+                sname,pdgid = ParseSpectraName(hname)
+                self.spectra[sname].append(pdgid, hist, path, sname)
+
+def CreateEmptyRebdsimFile(outputfilename, nOriginalEvents=1):
+    """
+    Create an empty rebdsim format file with the layout of folders.
+    Returns the ROOT.TFile object.
+    """
+    if not outputfilename.endswith(".root"):
+        outputfilename += ".root"
+
+    dc = _ROOT.DataDummyClass()
+    f = dc.CreateEmptyRebdsimFile(outputfilename, nOriginalEvents)
+    return f
+
+def WriteROOTHistogramsToDirectory(tfile, directoryName, histograms):
+    """
+    :param tfile: TFile object to write to.
+    :type  tfile: ROOT.TFile.
+    :param directoryName: Full path of directory you wish to write the histograms to.
+    :type  directoryName: str  (e.g. "Event/PerEntryHistograms" )
+    :param histograms:  List of ROOT histograms to write.
+    :type  histograms: [ROOT.TH1,..]
+    
+    Write a list of hitograms (ROOT.TH*) to a directory (str) in a ROOT.TFile instance.
+    """
+    tfile.cd(directoryName)
+    directory = tfile.Get(directoryName)
+    for hist in histograms:
+        directory.WriteObject(hist, hist.GetName())
+    
+            
 class BDSAsciiData(list):
     """
     General class representing simple 2 column data.
@@ -748,6 +822,7 @@ class TH1(ROOTHist):
         self.xcentres   = _np.zeros(self.nbinsx)
         self.xlowedge   = _np.zeros(self.nbinsx)
         self.xhighedge  = _np.zeros(self.nbinsx)
+        self.xrange     = (0,0)
 
         # data holders
         self.contents  = _np.zeros(self.nbinsx)
@@ -761,9 +836,12 @@ class TH1(ROOTHist):
             self.xlowedge[i]  = xaxis.GetBinLowEdge(i+1)
             self.xhighedge[i] = self.xlowedge[i] + self.xwidths[i]
             self.xcentres[i]  = xaxis.GetBinCenter(i+1)
+        self.xrange = (self.xlowedge[0],self.xhighedge[-1])
 
         if extractData:
             self._GetContents()
+
+        self.integral = _np.sum(self.contents)
 
     def _GetContents(self):
         for i in range(self.nbinsx):
@@ -784,6 +862,7 @@ class TH2(TH1):
         self.ycentres  = _np.zeros(self.nbinsy)
         self.ylowedge  = _np.zeros(self.nbinsy)
         self.yhighedge = _np.zeros(self.nbinsy)
+        self.yrange    = (0,0)
 
         self.contents = _np.zeros((self.nbinsx,self.nbinsy))
         self.errors   = _np.zeros((self.nbinsx,self.nbinsy))
@@ -794,9 +873,12 @@ class TH2(TH1):
             self.ylowedge[i]  = yaxis.GetBinLowEdge(i+1)
             self.yhighedge[i] = self.ylowedge[i] + self.ywidths[i]
             self.ycentres[i]  = yaxis.GetBinCenter(i+1)
+        self.yrange = (self.ylowedge[0],self.yhighedge[-1])
 
         if extractData:
             self._GetContents()
+
+        self.integral = _np.sum(self.contents)
 
     def _GetContents(self):
         for i in range(self.nbinsx) :
@@ -819,6 +901,7 @@ class TH3(TH2):
         self.zcentres  = _np.zeros(self.nbinsz)
         self.zlowedge  = _np.zeros(self.nbinsz)
         self.zhighedge = _np.zeros(self.nbinsz)
+        self.zrange    = (0,0)
 
         self.contents = _np.zeros((self.nbinsx,self.nbinsy,self.nbinsz))
         self.errors   = _np.zeros((self.nbinsx,self.nbinsy,self.nbinsz))
@@ -829,9 +912,12 @@ class TH3(TH2):
             self.zlowedge[i]  = zaxis.GetBinLowEdge(i+1)
             self.zhighedge[i] = self.zlowedge[i] + self.zwidths[i]
             self.zcentres[i]  = zaxis.GetBinCenter(i+1)
+        self.zrange = (self.zlowedge[0],self.zhighedge[-1])
 
         if extractData:
             self._GetContents()
+
+        self.integral = _np.sum(self.contents)
 
     def _GetContents(self):
         for i in range(self.nbinsx):
@@ -917,7 +1003,7 @@ class _SamplerData(object):
 
 class PhaseSpaceData(_SamplerData):
     """
-    Pull phase space data from a loaded DataLoader instance of raw data.
+    Pull phase space data from a loaded DataLoader instance of raw data for all events.
 
     Extracts only: 'x','xp','y','yp','z','zp','energy','T'
 
@@ -937,7 +1023,7 @@ class PhaseSpaceData(_SamplerData):
 
 class SamplerData(_SamplerData):
     """
-    Pull sampler data from a loaded DataLoader instance of raw data.
+    Pull sampler data from a loaded DataLoader instance of raw data for all events.
 
     Loads all data in a given sampler.
 
@@ -1293,6 +1379,8 @@ class ModelData(object):
         return cls(data)
 
     def _getData(self, interface, rootobj):
+        # remove when fixed this
+        _np.warnings.filterwarnings('ignore', category=_np.VisibleDeprecationWarning)
         for name in interface:
             try:
                 setattr(self, name, _np.array(getattr(rootobj, name)))
@@ -1305,6 +1393,18 @@ class ModelData(object):
                     pass # just ignore it
             except ValueError:
                 pass # just ignore it
+
+        possibleDicts = ["materialIDToName", "materialNameToID"]
+        for n in possibleDicts:
+            if hasattr(self, n):
+                try:
+                    setattr(self, n, dict(getattr(self,n)))
+                except ValueError:
+                    pass
+        if hasattr(self, 'materialIDToName'):
+            self.materialIDToName = {int(k):v for k,v in self.materialIDToName.items()}
+        if hasattr(self, 'materialNameToID'):
+            self.materialNameToID = {k:int(v) for k,v in self.materialNameToID.items()}
 
     def GetApertureData(self, removeZeroLength=False, removeZeroApertures=True, lengthTolerance=1e-6):
         """