+++ docs/source/version_history.rst
@@ -10,6 +10,17 @@ New Features
 
 * Create a nice copy of the header information from any (re)bdsim file when
   loading with pybdsim using only Python types.
+* New ratio plot for 2x 1D histograms. See `pybdsim.Plot.Histogram1DRatio`.
+* pybdsim.Data.TH1,2,3 now have :code:`xrange`, :code:`yrange`, and :code:`zrange` members
+  where approriate with a conveninent tuple of the range in each dimension.
+* Field plotting functions now tolerate Field class objects as well as filenames to make
+  it easier to check field objects as you're making them.
+* New field plotting for 2D field maps showing each component.
+* New field reflection utility function `pybdsim.Field.MirrorDipoleQuadrant1` for 2D fields.
+* New field plotting function `pybdsim.Field.Plot2DXYConnectionOrder` to see the order
+  an array is written in. This can be used to validate any field manipulations.
+* New field plotting function `pybdsim.Field.Plot1DFxFyFz` to see field components in 1D.
+* Field loading automatically works for dimensions such as X, Z for 2D instead of X, Y now.
 
 Bug Fixes
 ---------
@@ -17,6 +28,15 @@ Bug Fixes
 * pybdsim would throw an exception that librebdsim and libbdsimRootEvent could not be
   loaded and stop if the libraries had been already loaded separately outside pybdsim.
   This has been fixed by fixing the interpretation of the error codes from ROOT.
+* Fix warning about "nonposy" in matplotlib version for log scales.
+* Fix check in Run of if it's a ROOT file or not. Simplify it to use file extension.
+
+General
+-------
+
+* Better automatic ranges for Histogram1DMultiple plots by default.
+* Better field loading in `pybdsim.Field.Load`. Returns the same Field object
+  from pybdsim as you would write.
 
 v2.4.0 - 2021 / 06 / 16
 =======================